+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 19.10
+
+* New query parameter has been added to the `GET /1.0/users/me/rooms/messages` endpoint:
+
+  * `distinct`
+
 ## 19.09
 
 * New state has been added to the `PUT /1.0./users/{user_uuid}/presence` endpoint: